+++ framework/bug-mining/initialize-revisions.pl
@@ -31,7 +31,7 @@ layout and perform a sanity check for each revision.
 
 =head1 SYNOPSIS
 
-initialize-revisions.pl -p project_id -w work_dir [ -b bug_id]
+initialize-revisions.pl -p project_id -w work_dir [-s subproject] [ -b bug_id] [-D]
 
 =head1 OPTIONS
 
@@ -45,10 +45,19 @@ The id of the project for which the meta data should be generated.
 
 The working directory used for the bug-mining process.
 
+=item B<-s F<subproject>>
+
+The subproject to be mined (if not the root directory)
+
 =item B<-b C<bug_id>>
 
 Only analyze this bug id. The bug_id has to follow the format B<(\d+)(:(\d+))?>.
-Per default all bug ids, listed in the commit-db, are considered.
+Per default all bug ids, listed in the active-bugs csv, are considered.
+
+=item B<-D>
+
+Debug: Enable verbose logging and do not delete the temporary check-out directory
+(optional).
 
 =back
 
@@ -67,13 +76,15 @@ use DB;
 use Utils;
 
 my %cmd_opts;
-getopts('p:b:w:', \%cmd_opts) or pod2usage(1);
+getopts('p:b:w:s:D', \%cmd_opts) or pod2usage(1);
 
 pod2usage(1) unless defined $cmd_opts{p} and defined $cmd_opts{w};
 
 my $PID = $cmd_opts{p};
 my $BID = $cmd_opts{b};
 my $WORK_DIR = abs_path($cmd_opts{w});
+my $SUBPROJ = $cmd_opts{s};
+$DEBUG = 1 if defined $cmd_opts{D};
 
 # Check format of target bug id
 if (defined $BID) {
@@ -120,6 +131,11 @@ sub _init_version {
     # minimized patch to obtain the buggy version.
     $project->{_vcs}->checkout_vid("${vid}", $work_dir) or die "Cannot checkout $vid version";
 
+    if (defined $SUBPROJ) {
+        $work_dir .= "/$SUBPROJ/";
+        $project->{prog_root} = $work_dir;
+    }
+
     system("mkdir -p $ANALYZER_OUTPUT/$bid");
     if (-e "$work_dir/build.xml") {
         my $cmd = " cd $work_dir" .
@@ -128,7 +144,7 @@ sub _init_version {
     } elsif (-e "$work_dir/pom.xml") {
         # Run maven-ant plugin and overwrite the original build.xml whenever a maven build file exists
         my $cmd = " cd $work_dir" .
-                  " && mvn ant:ant -Doverwrite=true 2>&1" .
+                  " && mvn ant:ant -Doverwrite=true 2>&1 -Dhttps.protocols=TLSv1.2" .
                   " && patch build.xml $PROJECT_DIR/build.xml.patch 2>&1" .
                   " && rm -rf $GEN_BUILDFILE_DIR/$rev_id && mkdir -p $GEN_BUILDFILE_DIR/$rev_id 2>&1" .
                   " && cp maven-build.* $GEN_BUILDFILE_DIR/$rev_id 2>&1" .
@@ -138,10 +154,14 @@ sub _init_version {
         $cmd = " cd $work_dir" .
                " && java -jar $LIB_DIR/analyzer.jar $work_dir $ANALYZER_OUTPUT/$bid maven-build.xml 2>&1";
         Utils::exec_cmd($cmd, "Run build-file analyzer on maven-ant.xml.") or die;
+	
+        # Fix broken dependency links
+        my $fix_dep = "cd $work_dir && sed \'s\/https:\\/\\/oss\\.sonatype\\.org\\/content\\/repositories\\/snapshots\\//http:\\/\\/central\\.maven\\.org\\/maven2\\/\/g\' maven-build.xml >> temp && mv temp maven-build.xml";
+        Utils::exec_cmd($fix_dep, "Fixing broken dependency links.");
 
         # Get dependencies if it is maven-ant project
         my $download_dep = "cd $work_dir && ant -Dmaven.repo.local=\"$PROJECT_DIR/lib\" get-deps";
-        Utils::exec_cmd($download_dep, "Download dependencies for maven-ant.xml");
+        Utils::exec_cmd($download_dep, "Download dependencies for maven-ant.xml.");
     } else {
         # TODO add support for other build systems
         die "Unsupported build system";
@@ -171,7 +191,7 @@ sub _bootstrap {
     $project->export_diff($v2, $v1, "$PATCH_DIR/$bid.test.patch", "$test_f");
 }
 
-my @ids = $project->get_version_ids();
+my @ids = $project->get_bug_ids();
 if (defined $BID) {
     if ($BID =~ /(\d+):(\d+)/) {
         @ids = grep { ($1 <= $_) && ($_ <= $2) } @ids;
@@ -208,4 +228,4 @@ print("\n--- Add the following to the <fileset> tag identified by the id 'all.ma
 system("cat $ANALYZER_OUTPUT/*/includes | sort -u | while read -r include; do echo \"<include name='\"\$include\"' />\"; done");
 system("cat $ANALYZER_OUTPUT/*/excludes | sort -u | while read -r exclude; do echo \"<exclude name='\"\$exclude\"' />\"; done");
 
-system("rm -rf $TMP_DIR");
+system("rm -rf $TMP_DIR") unless $DEBUG;
